[ruby/rubygems] Omit rubygems tests that fail on JRuby + Windows
Extend symlink_supported? to probe readlink so the symlink-based
tests omit cleanly when JRuby on Windows lacks readlink, switch
test_gem_paths_do_not_contain_symlinks to use the helper, and add
per-test omits for the remaining JRuby/Windows incompatibilities
(install permissions, project sanity, require/realworld, open
command, git submodule checkout, operating-system defaults).
test_self_load_escape_* now also rescues Errno::EACCES which JRuby
on Windows raises instead of Errno::EINVAL.

https://github.com/ruby/rubygems/commit/321782d903

+++ test/rubygems/helper.rb
@@ -1268,11 +1268,13 @@ Also, a list:
     if @@symlink_supported.nil?
       begin
         File.symlink(File.join(@tempdir, "a"), File.join(@tempdir, "b"))
+        File.readlink(File.join(@tempdir, "b"))
       rescue NotImplementedError, SystemCallError
         @@symlink_supported = false
       else
-        File.unlink(File.join(@tempdir, "b"))
         @@symlink_supported = true
+      ensure
+        File.unlink(File.join(@tempdir, "b")) if File.symlink?(File.join(@tempdir, "b"))
       end
     end
     @@symlink_supported
